Fix webserver tests
- Refactor some test classes to only have @Template annotation
  on the most down class in the hierarchy. This avoids ce-arq
  to instantiate multiple templates.
- Fix testrunner pod image tag and readiness probe
- Enable webserver module in parent pom
