+++ webserver/src/test/java/org/jboss/test/arquillian/ce/webserver/WebServerTomcat8BasicSecureTest.java
@@ -23,15 +23,18 @@
 
 package org.jboss.test.arquillian.ce.webserver;
 
+import java.net.URL;
+
+import javax.websocket.ClientEndpoint;
+
+import org.jboss.arquillian.ce.api.OpenShiftResource;
+import org.jboss.arquillian.ce.api.OpenShiftResources;
 import org.jboss.arquillian.ce.api.Template;
 import org.jboss.arquillian.ce.cube.RouteURL;
 import org.jboss.arquillian.container.test.api.RunAsClient;
 import org.jboss.arquillian.junit.Arquillian;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-
-import javax.websocket.ClientEndpoint;
-import java.net.URL;
 
 /**
  * @author fspolti
@@ -40,8 +43,11 @@ import java.net.URL;
 @Template(url = "https://raw.githubusercontent.com/${template.repository:jboss-openshift}/application-templates/${template.branch:master}/webserver/jws30-tomcat8-https-s2i.json",
         labels = "application=jws-app"
 )
+@OpenShiftResources({
+    @OpenShiftResource("https://raw.githubusercontent.com/${template.repository:jboss-openshift}/application-templates/${template.branch:master}/secrets/jws-app-secret.json")
+})
 @ClientEndpoint
-public class WebServerTomcat8BasicSecureTest extends WebServerTomcat8BasicTest {
+public class WebServerTomcat8BasicSecureTest extends WebserverTestBase {
 
     @Test
     @RunAsClient
