+++ webserver/src/test/java/org/jboss/test/arquillian/ce/webserver/WebServerTomcat8MySQLDbPVTest.java
@@ -35,6 +35,6 @@ import org.junit.runner.RunWith;
 @Template(url = "https://raw.githubusercontent.com/${template.repository:jboss-openshift}/application-templates/${template.branch:master}/webserver/jws30-tomcat8-mysql-persistent-s2i.json",
         labels = "application=jws-app"
 )
-public class WebServerTomcat8MySQLDbPVTest extends WebServerTomcat7MySQLDbPVTest {
+public class WebServerTomcat8MySQLDbPVTest extends WebServerTomcatMySQLDbPVTestBase {
 
 }
